Isolate common types for Astarte
